# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tender API",
    "version": "0.1.0",
    "description": "Any-chain crypto checkout. Buyers pay from any chain; merchants settle in one asset on Monad.\n\nMerchant routes need **Authorize → `tk_live_…`** (create one with `npm run merchant:create`). Public routes are what the buyer checkout calls. Money is always a decimal **string**; the wire is **snake_case**."
  },
  "components": {
    "securitySchemes": {
      "merchantKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Merchant API key (tk_live_…)"
      },
      "platformKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dashboard platform key (tp_…). On /v1 routes, also send `X-Tender-Merchant: mer_…`. Server-side only."
      }
    }
  },
...
  "paths": {
...
    "/v1/invoices/{id}": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "summary": "One invoice, with its payments",
        "security": [
          {
            "merchantKey": []
          },
          {
            "platformKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "DETECTED",
                        "SETTLED",
                        "OVERPAID",
                        "UNDERPAID",
                        "EXPIRED",
                        "CANCELLED",
                        "NEEDS_RECOVERY"
                      ]
                    },
                    "amount_expected": {
                      "type": "string",
                      "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "expires_at": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "redirect_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "metadata": {
                      "anyOf": [
                        {
                          "type": "object",
                          "propertyNames": {
                            "type": "string"
                          },
                          "additionalProperties": {}
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "addresses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "chain": {
                            "type": "string",
                            "minLength": 1
                          },
                          "address": {
                            "type": "string"
                          },
                          "minimum": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                          }
                        },
                        "required": [
                          "chain",
                          "address"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "payments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "invoice_id": {
                            "type": "string"
                          },
                          "tx_hash": {
                            "type": "string"
                          },
                          "from_chain": {
                            "type": "string",
                            "minLength": 1
                          },
                          "amount_in": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                          },
                          "amount_settled": {
                            "anyOf": [
                              {
                                "type": "string",
                                "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "sender": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "refunded_amount": {
                            "anyOf": [
                              {
                                "type": "string",
                                "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "DETECTED",
                              "SETTLED",
                              "FAILED",
                              "REFUNDED"
                            ]
                          },
                          "first_seen_at": {
                            "type": "string"
                          },
                          "settled_at": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "invoice_id",
                          "tx_hash",
                          "from_chain",
                          "amount_in",
                          "status",
                          "first_seen_at"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "token",
                    "status",
                    "amount_expected",
                    "currency",
                    "reference",
                    "expires_at",
                    "created_at",
                    "addresses"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/invoices/{id}/cancel": {
      "post": {
        "tags": [
          "Invoices"
        ],
        "summary": "Cancel a pending invoice",
        "security": [
          {
            "merchantKey": []
          },
          {
            "platformKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "DETECTED",
                        "SETTLED",
                        "OVERPAID",
                        "UNDERPAID",
                        "EXPIRED",
                        "CANCELLED",
                        "NEEDS_RECOVERY"
                      ]
                    },
                    "amount_expected": {
                      "type": "string",
                      "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "expires_at": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "redirect_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "metadata": {
                      "anyOf": [
                        {
                          "type": "object",
                          "propertyNames": {
                            "type": "string"
                          },
                          "additionalProperties": {}
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "addresses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "chain": {
                            "type": "string",
                            "minLength": 1
                          },
                          "address": {
                            "type": "string"
                          },
                          "minimum": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                          }
                        },
                        "required": [
                          "chain",
                          "address"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "payments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "invoice_id": {
                            "type": "string"
                          },
                          "tx_hash": {
                            "type": "string"
                          },
                          "from_chain": {
                            "type": "string",
                            "minLength": 1
                          },
                          "amount_in": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                          },
                          "amount_settled": {
                            "anyOf": [
                              {
                                "type": "string",
                                "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "sender": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "refunded_amount": {
                            "anyOf": [
                              {
                                "type": "string",
                                "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "DETECTED",
                              "SETTLED",
                              "FAILED",
                              "REFUNDED"
                            ]
                          },
                          "first_seen_at": {
                            "type": "string"
                          },
                          "settled_at": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "invoice_id",
                          "tx_hash",
                          "from_chain",
                          "amount_in",
                          "status",
                          "first_seen_at"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "token",
                    "status",
                    "amount_expected",
                    "currency",
                    "reference",
                    "expires_at",
                    "created_at",
                    "addresses"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}